initial commit with data and interfaces
