--- src/App.scss
@@ -1,60 +1,0 @@
-.root {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 2rem;
-  display: flex;
-  text-align: center;
-  &__box {
-    flex: 1 0 200px;
-    padding: 1rem;
-    text-align: left;
-  }
-  &__box:nth-child(2) {
-    flex: 1 0 600px;
-    p {
-      text-align: justify;
-      margin-top: 5px;
-      margin-bottom: 30px;
-    }
-  }
-  &__photoName {
-    display: flex;
-    align-items: end;
-    gap: 10px;
-    img {
-      height: 70px;
-    }
-    p {
-      margin: 0;
-      line-height: 1;
-    }
-  }
-}
-
-// .logo {
-//   height: 6em;
-//   padding: 1.5em;
-//   will-change: filter;
-//   transition: filter 300ms;
-// }
-// .logo:hover {
-//   filter: drop-shadow(0 0 2em #646cffaa);
-// }
-// .logo.react:hover {
-//   filter: drop-shadow(0 0 2em #61dafbaa);
-// }
-
-// @keyframes logo-spin {
-//   from {
-//     transform: rotate(0deg);
-//   }
-//   to {
-//     transform: rotate(360deg);
-//   }
-// }
-
-// @media (prefers-reduced-motion: no-preference) {
-//   a:nth-of-type(2) .logo {
-//     animation: logo-spin infinite 20s linear;
-//   }
-// }
